+++ CHANGES.md
@@ -94,6 +94,8 @@
 
 [develop changelog]
 
+* Fix API typo and a regression 'archivefirstmatch'
+
 
 ### 0.14.1 (2018-02-03 22:40:00 UTC)